requirements.txt
Django~=5.0.6
dj-database-url==0.5.0
psycopg2-binary==2.9.1
djangorestframework~=3.15.2